Replace octet reading ASN1 code with main parser for simplicity
