+++ gradle/scripts/detekt-config.yml
@@ -1,24 +1,33 @@
 build:
-  maxIssues: 0
+    maxIssues: 0
 
 formatting:
-  MaximumLineLength:
-    active: false
+    ArgumentListWrapping:
+        maxLineLength: 145 # increased from 120
+    ContextReceiverMapping:
+        maxLineLength: 145 # increased from 120
+    FunctionSignature:
+        maxLineLength: 145 # increased from 120
+    MaximumLineLength:
+        active: false
+        maxLineLength: 145 # increased from 120
+    ParameterListWrapping:
+        maxLineLength: 145 # increased from 120
 
 style:
-  MaxLineLength:
-    maxLineLength: 145 # increased from 120
-  MagicNumber:
-    excludes: ["**/test/**","**/androidTest/**","**Test.kt","**Spec.kt","**Spek.kt","**.kts"]
+    MaxLineLength:
+        maxLineLength: 145 # increased from 120
+    MagicNumber:
+        excludes: [ "**/test/**","**/androidTest/**","**Test.kt","**Spec.kt","**Spek.kt","**.kts" ]
 
 complexity:
-  LongMethod:
-    excludes: "**/test/**,**/androidTest/**,**/*.Test.kt,**/*.Spec.kt,**/*.Spek.kt"
+    LongMethod:
+        excludes: "**/test/**,**/androidTest/**,**/*.Test.kt,**/*.Spec.kt,**/*.Spek.kt"
 
 exceptions:
-  TooGenericExceptionThrown:
-    excludes: ['**/test/**', '**/androidTest/**', '**/commonTest/**', '**/jvmTest/**', '**/jsTest/**', '**/iosTest/**']
+    TooGenericExceptionThrown:
+        excludes: [ '**/test/**', '**/androidTest/**', '**/commonTest/**', '**/jvmTest/**', '**/jsTest/**', '**/iosTest/**' ]
 
 naming:
-  FunctionNaming:
-    ignoreAnnotated: ['Composable']
+    FunctionNaming:
+        ignoreAnnotated: [ 'Composable' ]
